Add tests that resize window
This is pretty non-trivial. Browsers will not let you resize the primary
window. However, Chrome let developers resize a window that was created
as a popup.

By opening Chrome with the `--disable-popup-blocking` flag and using
`window.open` in the acceptance tests it is possible (but hacky) to run
assertions against the app while changing the window size.